requirements.txt
# Base
environs==9.5.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
 
# Internationalization
django-modeltranslation==1.3.0
django-rosetta==0.9.9
 
# Development
django-debug-toolbar==4.2.0
django-extensions==3.2.3
 
# Production
gunicorn==21.2.0
whitenoise==6.6.0
 
# Database
django-environ==0.11.2
 
# Security
django-cors-headers==4.3.1
 
# Testing
coverage==7.3.2
pytest==7.4.3
pytest-django==4.7.0
 
# Code quality
black==23.11.0
flake8==6.1.0
isort==5.12.0
 
# Documentation
Sphinx==7.2.6
 
# Images
Pillow==10.1.0
 
# Optional
django-ckeditor==6.7.0  #